review: fix all confirmed findings from the 0.3.0 adversarial review pass
- SizeLimitExceeded salvage gated on a caller-set SizeLimit: a server's own
  administrative limit (OpenLDAP sizelimit 500) cutting an unbounded stream is
  a loud LimitsExceeded error again, not a clean-looking truncated result.
- -ResultSetSize truncation warning extended to Search-ADxAccount and the
  membership cmdlets (the README claimed all cmdlets warned; only the object
  presets did), and the +1 probe uses long arithmetic so [int]::MaxValue
  cannot wrap into a negative page size crash.
- -approx on GeneralizedTime attributes shares -eq's sub-second doctrine
  instead of silently truncating; a sub-second -ge/-gt bound in DateTime's
  final second is a clean translation error instead of an unwrapped
  ArgumentOutOfRangeException; the wildcard-rejection message names
  exact-match operators generically instead of claiming -eq/-ne.
- -RecursiveMatch gated to link-valued DN attributes (member/memberOf/manager/
  managedBy/backlinks); the Dn-syntax gate had admitted objectCategory and
  distinguishedName, where the 1941 chain walk is degenerate.
- Get-ADxPrincipalGroupMembership completes the memberOf range walk before its
  foreign-partition warning; Get-ADxGroupNested opts out of primary-group
  warnings its query cannot be affected by.
- README GC bullet narrowed to Get-ADxGroupMember (the principal cmdlet
  matches by full SID and is immune); locked-out captions in example 27 and
  the Search-ADxAccount help aligned with the stale-lockout semantics; help
  XML regenerated.

+++ src/ADx.Cmdlets/Base/ADxObjectCmdletBase.cs
@@ -132,9 +132,11 @@ public abstract class ADxObjectCmdletBase : ADxCmdletBase
         // against 31ms once the page is capped, on the same query returning the same object.
         // Only ever shrinks the page, so bulk enumeration (ResultSetSize 0 = unlimited) is
         // untouched. The +1 is the truncation probe: one extra entry proves whether the cap
-        // cut the result set, and it rides inside the same page.
+        // cut the result set, and it rides inside the same page. Long arithmetic: with
+        // -ResultSetSize [int]::MaxValue an Int32 +1 wraps negative and the page control
+        // throws.
         var effectivePageSize = ResultSetSize > 0
-            ? Math.Min(ResultPageSize, ResultSetSize + 1)
+            ? (int)Math.Min(ResultPageSize, (long)ResultSetSize + 1)
             : ResultPageSize;
 
         var spec = new LdapSearchSpec(
@@ -151,7 +153,7 @@ public abstract class ADxObjectCmdletBase : ADxCmdletBase
         var enumerable = iterator.StreamAsync(
             spec,
             // One past the cap: the extra entry only proves truncation and is never emitted.
-            maxItems: ResultSetSize > 0 ? ResultSetSize + 1 : 0,
+            maxItems: ResultSetSize > 0 ? (long)ResultSetSize + 1 : 0,
             onPageComplete: info => EnqueueVerbose(
                 $"Page {info.PageIndex}: {info.EntriesInPage} entries ({info.TotalEmitted} total)."),
             skipFirst: 0,
